Update key-policy.md
Added text from the "five things you need to know" document before the membership/key info, modified the text in an attempt to reflect discussion at the town hall mtg and other feedback i've received.  i think the statement about suggested rates is vague enough that if we want to modify the info on the website and/or provide other coworking space rates as a comparison we can do that and this language will still apply.  
Attempted to make it clear in the "keys to the castle" section that the orientation/signature steps should happen regardless of whether or not you want a key.
Added people's last names where i knew them, and a line for people sign next to the things to learn (probably a better way to format this but i don't know how)
Combined Moneys and Anonymous Donations into one category (if people are specifically interested in gratipay or more info about the donation model, i will direct them to Chad and Elliott)
Questions:
1. Do we want to link specific individuals to the "things to learn" list, or just allow people to put two and two together? 
2. Can the "Talk to... about..." section be formatted a little more like a table? I don't know how to do this within github.